requirements.txt
paho-mqtt
influxdb-client
schedule
pyyaml
protobuf==5.27.0